# docs/swagger.yaml
basePath: /api
definitions:
  dto.AuthConfirmVerifyRequest:
    properties:
      address:
        example: "+77012345678"
        type: string
      code:
        maxLength: 6
        minLength: 6
        type: string
      type:
        description: UserID  int64  `json:"user_id" validate:"required" swaggertype:"integer"
          example:"1"`
        example: phone
        type: string
    required:
    - address
    - code
    - type
    type: object
  dto.AuthHelloResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      role_name:
        type: string
    type: object
  dto.AuthLoginRequest:
    properties:
      email:
        example: test@mail.com
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        example: "+77012345678"
        type: string
    required:
    - password
    type: object
  dto.AuthLoginResponse:
    properties:
      access_token:
        type: string
      id:
        type: integer
      name:
        type: string
      refresh_token:
        type: string
      role_name:
        type: string
    type: object
  dto.AuthRegisterRequest:
    properties:
      confirm_type:
        example: phone or email
        type: string
      email:
        example: test@mail.com
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        example: "77012345678"
        type: string
    required:
    - confirm_type
    - name
    - password
    type: object
  dto.AuthRegisterResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      otp_expires_at:
        type: string
      role_name:
        type: string
    type: object
  dto.AuthSendVerifyRequest:
    properties:
      address:
        example: "+77012345678"
        type: string
      type:
        example: phone
        type: string
    required:
    - address
    - type
    type: object
  dto.AuthSendVerifyResponse:
    properties:
      otp_expires_at:
        type: string
    type: object
  dto.AuthSession:
    properties:
      created_at:
        type: string
      ip:
        type: string
      jti:
        type: string
      role_id:
        type: integer
      user_agent:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
      user_phone_number:
        type: string
    type: object
  dto.AuthSessionResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/dto.AuthSession'
        type: array
    type: object
  dto.AuthUpdatePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - new_password
    - old_password
    - user_id
    type: object
  dto.UserResponse:
    properties:
      email:
        example: almas@gmail.com
        type: string
      id:
        type: integer
      name:
        example: almas
        type: string
      phone_number:
        example: "+77012345678"
        type: string
      role_name:
        example: user
        type: string
    type: object
  dto.UsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
host: localhost:3199
info:
  contact: {}
  title: github.com/AlmasNurbayev/go_fiber_boilerplate API documentation
  version: 1.0.0
paths:
  /auth/confirm-verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthConfirmVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
      summary: Send verify code to user address
      tags:
      - Auth
  /auth/hello:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthHelloResponse'
        "401":
          description: authentication failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Check auth token
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthLoginResponse'
        "401":
          description: authentication failed
          schema:
            type: string
      summary: Login as user, returns access and refresh tokens
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthLoginResponse'
        "401":
          description: authentication failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Check refresh token, returns new access and refresh tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthRegisterResponse'
        "400":
          description: 'Key: ''AuthRegisterRequest.Password'' Error:Field validation
            for ''Password'' failed on the ''min'' tag'
          schema:
            type: string
        "409":
          description: 'значение для поля уже существует (ограничение уникальности:
            users_XXXX_key)'
          schema:
            type: string
      summary: Register as user
      tags:
      - Auth
  /auth/send-verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthSendVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthSendVerifyResponse'
        "401":
          description: authentication failed
          schema:
            type: string
      summary: Verify user addresses
      tags:
      - Auth
  /auth/sessions/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthSessionResponse'
        "401":
          description: authentication failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get sessions by user id, only user can get his own sessions
      tags:
      - Auth
  /auth/sessions/{jti}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Session jti
        in: path
        name: jti
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "401":
          description: authentication failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Revoke session by jti, only user can revoke his own session
      tags:
      - Auth
  /auth/update-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AuthUpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: authentication failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user password for authenticated user
      tags:
      - Auth
  /users/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UsersResponse'
        "401":
          description: authentication failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Check auth token
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"